requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
anthropic>=0.40.0
httpx>=0.26.0
sse-starlette>=1.6.0
reportlab>=4.0.0
python-docx>=1.0.0
python-pptx>=0.6.23
redis>=5.0.1
python-jose[cryptography]==3.3.0
structlog>=24.1.0
